Moved NIC main into examples
